feat(ui): unify add/remove button styles across all resume sections
