+++ src/modules/resume/components/form/resume-form.tsx
@@ -2,8 +2,8 @@ import { useFormContext } from "react-hook-form";
 import { ResumeFormValues } from "@resume/schemas/resume.schema";
 import SectionProfile from "@resume/components/form/section-profile";
 import SectionExperience from "@resume/components/form/section-experience";
-import SectionEducation from "@resume/components/form/section-education-new";
-import SectionProject from "@resume/components/form/section-project-new";
+import SectionEducation from "@resume/components/form/section-education";
+import SectionProject from "@resume/components/form/section-project";
 import SectionSkills from "@resume/components/form/section-skills";
 import SectionCustomFields from "@resume/components/form/section-custom-fields";
